Actually implement 0MQ HWM setting

+++ src/redis_zmq.h
@@ -5,7 +5,7 @@
 
 extern unsigned int redis_zmq_num_endpoints;
 extern char **redis_zmq_endpoints;
-extern unsigned int redis_zmq_hwm;
+extern uint64_t redis_zmq_hwm;
 
 void dispatchExpiryMessage(redisDb *db, robj *key);
